@@ -16,7 +16,8 @@
     "@angular/common": "^6.0.3",
     "@angular/compiler": "^6.0.3",
     "@angular/core": "^6.0.3",
-    "@angular/flex-layout": "^6.0.0-beta.17",
+    "@angular/flex-layout": "^5.0.0-beta.14",
+    "@angular/flex-layout": "^6.0.0-beta.15",
     "@angular/forms": "^6.0.3",
     "@angular/http": "^6.0.3",
     "@angular/material": "^6.4.2",
@@ -30,18 +31,22 @@
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
+    "modules": "^0.4.0",
+    "ng": "0.0.0",
     "ng-modules": "^2.0.9",
     "ngx-bootstrap": "^3.0.1",
+    "node": "^10.9.0",
     "node-module": "^2.0.0",
     "node-modules": "^1.0.1",
     "popper.js": "^1.14.4",
-    "rxjs": "^6.0.0",
+    "rxjs": "^6.3.0",
+    "rxjs-compat": "^6.0.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angular/compiler-cli": "^6.0.3",
     "@angular-devkit/build-angular": "~0.6.8",
-    "typescript": "~2.7.2",
+    "typescript": "~2.9.0",
     "@angular/cli": "~6.0.8",
     "@angular/language-service": "^6.0.3",
     "@types/jasmine": "~2.8.6",
