@@ -5,19 +5,23 @@ import { RegistrationComponent } from './registration/registration.component';
 import { LoginComponent } from './login/login.component';
 import { AuthGuard } from './guards/auth.guard';
 import { PortfoliosComponent } from './portfolios/portfolios.component';
-import { VisitingCardComponent } from './portfolios/visiting-card/visiting-card.component';
 import { HomeComponent } from './home/home.component';
+import { StartComponent } from './start/start.component';
+import { AboutUsComponent } from './about-us/about-us.component';
+import { HelpComponent } from './help/help.component';
+import { ContactComponent } from './contact/contact.component';
 
 const routes: Routes = [
   {path: 'home', component: HomeComponent},
   {path: 'profile', component: ProfileComponent, canActivate: [AuthGuard]},
   {path: 'portfolios', component: PortfoliosComponent, canActivate: [AuthGuard]},
   {path: 'register', component: RegistrationComponent},
-  {path: '', component: LoginComponent},
-  {path: 'portfolios/visiting', component: VisitingCardComponent, canActivate: [AuthGuard]}
-
-];
-
+  {path: 'login', component: LoginComponent},
+  {path: '', component: StartComponent},
+  {path: 'about', component: AboutUsComponent},
+  {path: 'help', component: HelpComponent},
+  {path: 'contact', component: ContactComponent}
+  ];
 @NgModule({
     imports: [
       RouterModule.forRoot(routes)
@@ -26,4 +30,4 @@ const routes: Routes = [
   })
 
   export class AppRoutingModule {}
-  export const routingComponents = [ProfileComponent, PortfoliosComponent, VisitingCardComponent, RegistrationComponent, LoginComponent];
+  export const routingComponents = [ProfileComponent, PortfoliosComponent, RegistrationComponent, LoginComponent, StartComponent, ContactComponent, HelpComponent, AboutUsComponent];
