@@ -5,28 +5,32 @@ import { Router } from '@angular/router';
 // import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
-  providers: [UserService]
+ selector: 'app-root',
+ templateUrl: './app.component.html',
+ styleUrls: ['./app.component.css'],
+ providers: [UserService]
 })
 export class AppComponent implements OnInit {
-  private user_check = false;
-  profile: String;
-  constructor(private authenticationService: AuthenticationService, private router: Router) {
+ private user_check = false;
+ profile: String;
+ constructor(private authenticationService: AuthenticationService, private router: Router) {
+   authenticationService.getLoggedInName.subscribe(() => {
+     console.log('hi');
+     this.ngOnInit();
+   });
+ }
+ ngOnInit() {
+ if (localStorage.getItem('currentUser') != null) {
+   this.user_check = true;
+   this.profile = JSON.parse(localStorage.getItem('currentUser'));
+ }
+ }
 
-  }
-  ngOnInit() {
-  if (localStorage.getItem('currentUser') != null) {
-    this.user_check = true;
-    this.profile = JSON.parse(localStorage.getItem('currentUser'));
-  }
-  }
-  logout() {
-    this.authenticationService.logout();
-    this.router.navigate(['/login']);
-    this.user_check = false;
-  }
+ logout() {
+   this.authenticationService.logout();
+   this.router.navigate(['/login']);
+   this.user_check = false;
+ }
 
 
 }
