@@ -2,9 +2,9 @@ package com.stackroute.downstreamservice.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-import lombok.Data;
+import org.springframework.data.mongodb.core.mapping.Document;
 
-@Data
+@Document
 @JsonIgnoreProperties
 public class Projects {
 
