@@ -1,30 +1,28 @@
 package com.stackroute.matchmaker.indexermodel;
+
 /**
  * 
  * @author syam
  *
  */
-//this is the model class , that is listned from kafka consumer
+// this is the model class , that is listned from kafka consumer
 public class ExperienceIndex {
 	private String profileId;
 	private String organizationName;
 	private String role;
-	private String startDate;
-	private String endDate;
+	private String duration;
 	private String message;
 
 	public ExperienceIndex() {
 		super();
 	}
 
-	public ExperienceIndex(String profileId, String organizationName, String role, String startDate, String endDate,
-			String message) {
+	public ExperienceIndex(String profileId, String organizationName, String role, String duration, String message) {
 		super();
 		this.profileId = profileId;
 		this.organizationName = organizationName;
 		this.role = role;
-		this.startDate = startDate;
-		this.endDate = endDate;
+		this.duration = duration;
 		this.message = message;
 	}
 
@@ -52,20 +50,12 @@ public class ExperienceIndex {
 		this.role = role;
 	}
 
-	public String getStartDate() {
-		return startDate;
+	public String getDuration() {
+		return duration;
 	}
 
-	public void setStartDate(String startDate) {
-		this.startDate = startDate;
-	}
-
-	public String getEndDate() {
-		return endDate;
-	}
-
-	public void setEndDate(String endDate) {
-		this.endDate = endDate;
+	public void setDuration(String duration) {
+		this.duration = duration;
 	}
 
 	public String getMessage() {
@@ -75,13 +65,16 @@ public class ExperienceIndex {
 	public void setMessage(String message) {
 		this.message = message;
 	}
-	
 
 	@Override
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
+		result = prime * result + ((duration == null) ? 0 : duration.hashCode());
 		result = prime * result + ((message == null) ? 0 : message.hashCode());
+		result = prime * result + ((organizationName == null) ? 0 : organizationName.hashCode());
+		result = prime * result + ((profileId == null) ? 0 : profileId.hashCode());
+		result = prime * result + ((role == null) ? 0 : role.hashCode());
 		return result;
 	}
 
@@ -94,19 +87,38 @@ public class ExperienceIndex {
 		if (getClass() != obj.getClass())
 			return false;
 		ExperienceIndex other = (ExperienceIndex) obj;
+		if (duration == null) {
+			if (other.duration != null)
+				return false;
+		} else if (!duration.equals(other.duration))
+			return false;
 		if (message == null) {
 			if (other.message != null)
 				return false;
 		} else if (!message.equals(other.message))
+			return false;
+		if (organizationName == null) {
+			if (other.organizationName != null)
+				return false;
+		} else if (!organizationName.equals(other.organizationName))
+			return false;
+		if (profileId == null) {
+			if (other.profileId != null)
+				return false;
+		} else if (!profileId.equals(other.profileId))
+			return false;
+		if (role == null) {
+			if (other.role != null)
+				return false;
+		} else if (!role.equals(other.role))
 			return false;
 		return true;
 	}
 
 	@Override
 	public String toString() {
-		return "ExperienceListner [profileId=" + profileId + ", organizationName=" + organizationName + ", role="
-				+ role + ", startDate=" + startDate + ", endDate=" + endDate  + ", message="
-				+ message + "]";
+		return "ExperienceIndex [profileId=" + profileId + ", organizationName=" + organizationName + ", role=" + role
+				+ ", duration=" + duration + ", message=" + message + "]";
 	}
 
 }
