@@ -7,18 +7,19 @@ import org.neo4j.ogm.annotation.StartNode;
 
 import com.stackroute.matchmaker.nodes.Company;
 import com.stackroute.matchmaker.nodes.ProfileId;
+
 /**
  * 
  * @author syam
  *
  */
-// this is the pojo class based on which the relationships and their properties in the neo4j  are created upon
+// this is the pojo class based on which the relationships and their properties
+// in the neo4j are created upon
 @RelationshipEntity(type = "employeeOf")
 public class EmployeeOfRelation {
 	@Id
 	private String role;
-	private String startDate;
-	private String endDate;
+	private String duration;
 	@StartNode
 	private ProfileId profileId;
 	@EndNode
@@ -28,11 +29,10 @@ public class EmployeeOfRelation {
 		super();
 	}
 
-	public EmployeeOfRelation(String role, String startDate, String endDate, ProfileId profileId, Company company) {
+	public EmployeeOfRelation(String role, String duration, ProfileId profileId, Company company) {
 		super();
 		this.role = role;
-		this.startDate = startDate;
-		this.endDate = endDate;
+		this.duration = duration;
 		this.profileId = profileId;
 		this.company = company;
 	}
@@ -45,20 +45,12 @@ public class EmployeeOfRelation {
 		this.role = role;
 	}
 
-	public String getStartDate() {
-		return startDate;
+	public String getDuration() {
+		return duration;
 	}
 
-	public void setStartDate(String startDate) {
-		this.startDate = startDate;
-	}
-
-	public String getEndDate() {
-		return endDate;
-	}
-
-	public void setEndDate(String endDate) {
-		this.endDate = endDate;
+	public void setDuration(String duration) {
+		this.duration = duration;
 	}
 
 	public ProfileId getProfileId() {
@@ -79,8 +71,8 @@ public class EmployeeOfRelation {
 
 	@Override
 	public String toString() {
-		return "EmployeeOfRelation [role=" + role + ", startDate=" + startDate + ", endDate=" + endDate + ", profileId="
-				+ profileId + ", company=" + company + "]";
+		return "EmployeeOfRelation [role=" + role + ", duration=" + duration + ", profileId=" + profileId + ", company="
+				+ company + "]";
 	}
 
 }
