@@ -2,11 +2,9 @@ package com.stackroute.matchmaker.nlpservice;
 
 import java.io.FileNotFoundException;
 
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.kafka.annotation.EnableKafka;
 
 /**
  * 
@@ -14,7 +12,7 @@ import org.springframework.kafka.annotation.EnableKafka;
  *
  */
 
-@EnableKafka
+
 @EnableDiscoveryClient
 @SpringBootApplication
 public class NlpServiceApplication {
