@@ -17,7 +17,7 @@ public class SwaggerConfig extends WebMvcConfigurationSupport {
    @Bean
    public Docket productApi() {
        return new Docket(DocumentationType.SWAGGER_2)
-               .select()                 .apis(RequestHandlerSelectors.basePackage("com.stackroute.movieapp.controller"))
+               .select()                 .apis(RequestHandlerSelectors.basePackage("com.stackroute.matchmaker.controller"))
                .paths(regex("/api.*"))
                .build();
            
