@@ -2,19 +2,29 @@ package com.stackroute.maverick.springbootzuulgatwayproxy;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 import org.springframework.context.annotation.Bean;
-
 import com.example.springbootzuulgatwayproxy.filters.ErrorFilter;
+import com.example.springbootzuulgatwayproxy.filters.JwtFilter;
 import com.example.springbootzuulgatwayproxy.filters.PostFilter;
 import com.example.springbootzuulgatwayproxy.filters.PreFilter;
 import com.example.springbootzuulgatwayproxy.filters.RouteFilter;
 
 @SpringBootApplication
-@EnableDiscoveryClient  // for eureka server to find this application
+@EnableDiscoveryClient // for eureka server to find this application
 @EnableZuulProxy // to act as api-gateway
 public class SpringBootZuulgatwayproxyApplication {
+
+	@Bean
+	public FilterRegistrationBean jwtFilter() {
+		final FilterRegistrationBean registrationBean = new FilterRegistrationBean();
+		registrationBean.setFilter(new JwtFilter());
+		registrationBean.addUrlPatterns("/downstream/*");
+		registrationBean.addUrlPatterns("/matchmaker/*");
+		return registrationBean;
+	}
 
 	public static void main(String[] args) {
 		SpringApplication.run(SpringBootZuulgatwayproxyApplication.class, args);
@@ -24,14 +34,17 @@ public class SpringBootZuulgatwayproxyApplication {
 	public PreFilter preFilter() {
 		return new PreFilter();
 	}
+
 	@Bean
 	public PostFilter postFilter() {
 		return new PostFilter();
 	}
+
 	@Bean
 	public ErrorFilter errorFilter() {
 		return new ErrorFilter();
 	}
+
 	@Bean
 	public RouteFilter routeFilter() {
 		return new RouteFilter();
